Use Ansible 2.3
Juniper has it's own repository for Ansible which has at least version 2.2.
This can be inferred from the use of the `systemd` module, which is
only available from Ansible 2.2.

There is also usage of the internal `ansible.inventory.Inventory` Python
class. This class no longer exists on this location as of version 2.4.